requirements.txt
# Epic FHIR Lab Reports - Python Dependencies

# HTTP requests for FHIR API calls
requests>=2.28.0

# JWT creation and validation for Epic authentication
PyJWT>=2.8.0

# Cryptography for RSA key handling
cryptography>=41.0.0

# Scheduling for 24-hour job runs
schedule>=1.2.0

# Environment variable management (optional)
python-dotenv>=1.0.0